+++ bracket-bevy/examples/roguelike_tutorial_4/main.rs
@@ -63,8 +63,8 @@ fn tick(
     draw_map(&map, &ctx);
     for (pos, render) in queries.p1().iter() {
         ctx.set(
-            pos.x as usize,
-            pos.y as usize,
+            pos.x,
+            pos.y,
             render.fg,
             render.bg,
             render.glyph,
